Let user pick first dive computer
For a dive with several dive computers the order of
the computers is a historical accident. But the first is special
as this is the one whose profile is displayed by default (and
in the mobile version, currently, only the first dive computer
is displayed).

This adds an entry to the View menu to make the current dive computer
the first one.

Signed-off-by: Robert C. Helling <helling@atdotde.de>