Correct BT detect for OSTC Sport
Apparently, OSTC Sport has a BT name like OSTCs<space><serial>.

Small code addition to detect this properly. As long as we
do not have an improved way of detection. Notice that most of
the HWs use the same BT hardware, so simple detection on offered
services will not work.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>